Test and Add Exception

+++ database/migrations/2020_02_19_081820_create_gallery_images_table.php
@@ -15,9 +15,9 @@ class CreateGalleryImagesTable extends Migration
     {
         Schema::create('gallery_images', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('page_id');
-            $table->integer('user_id');
-            $table->integer('album_id');
+            $table->unsignedBigInteger('page_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('album_id');
             $table->binary('photo');
             $table->timestamps();
         });
